requirements.txt
attrs==21.4.0
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pydantic==1.9.0
pyparsing==3.0.7
pytest==6.2.5
python-dotenv==0.19.2
toml==0.10.2
typing_extensions==4.0.1
